+++ src/Tests/if_subscription_does_not_exist_and_table_is_empty.cs
@@ -8,38 +8,36 @@ using Xunit.Abstractions;
 namespace Tests;
 
 // ReSharper disable once InconsistentNaming
-public class When_Subscription_Does_Not_Exist_And_Table_Is_Not_Empty(ITestOutputHelper testOutputHelper): DatabaseFixture(testOutputHelper)
+public class if_subscription_does_not_exist_and_table_is_empty(ITestOutputHelper testOutputHelper): DatabaseFixture(testOutputHelper)
 {
     [Fact]
-    public async Task Read_from_table_using_named_transaction_snapshot()
+    public async Task read_from_table_using_named_transaction_snapshot()
     {
         var ct = TimeoutTokenSource().Token;
-        var sharedNamingPolicy = new AttributeNamingPolicy();
+
         var connectionString = Container.GetConnectionString();
         var eventsTable = await CreateOutboxTable(NpgsqlDataSource.Create(connectionString), ct);
-
+        var sharedNamingPolicy = new AttributeNamingPolicy();
         var resolver = new OptionsBuilder()
             .JsonContext(PublisherContext.Default)
             .NamingPolicy(sharedNamingPolicy)
             .WithTable(o => o.Name(eventsTable))
             .Build();
-
         //subscriber ignored msg
-        await MessageAppender.AppendAsync( new PublisherUserDeleted(Guid.NewGuid(), Guid.NewGuid().ToString()), resolver, connectionString, ct);
+        await MessageAppender.AppendAsync(new PublisherUserDeleted(Guid.NewGuid(), Guid.NewGuid().ToString()), resolver, connectionString, ct);
 
         //poison message
         await InsertPoisoningMessage(connectionString, eventsTable, ct);
 
         var @event = new PublisherUserCreated(Guid.NewGuid(), Guid.NewGuid().ToString());
-        await MessageAppender.AppendAsync(@event, resolver, connectionString, ct);
-
         var @expected = new SubscriberUserCreated(@event.Id, @event.Name);
 
-        var ( _, subscriptionOptions) =
-            SetupFor<SubscriberUserCreated>(connectionString, eventsTable, SubscriberContext.Default, sharedNamingPolicy, Output.WriteLine);
+        await MessageAppender.AppendAsync(@event, resolver, connectionString, ct);
+
+        var ( _, subscriptionOptions) = SetupFor<SubscriberUserCreated>(connectionString, eventsTable,
+            SubscriberContext.Default, sharedNamingPolicy, Output.WriteLine);
         var subscription = new Subscription();
         await using var subscription1 = subscription.ConfigureAwait(false);
-
         await foreach (var envelope in subscription.Subscribe(_ => subscriptionOptions, ct).ConfigureAwait(false))
         {
             Assert.Equal(@expected, ((OkEnvelope)envelope).Value);
